Deleted old accordion.tsx and moved input.tsx to components/ui. Fixed prettier errors in 'npm run build' (#46)
